drop AstHtmlVisitor
